fix: comma in qty and float in presence

+++ database/migrations/2024_11_30_111835_create_presences_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->integer('month');
             $table->integer('year');
             $table->json('attendance_data');
-            $table->integer('jours')->nullable();
+            $table->float('jours')->nullable();
             $table->timestamps();
             $table->unique(['employe_id', 'month', 'year']);
         });
